Decompose statistics and ICLabel closeout modules
